Fix actual type errors (#520)
* Make typechecker happy in configs (+sami)

* Fix types in utils

* Fix type issues in training

* Fix orchestrator types

* Fix infer types

* Remove type ignores

* Revert some changes

* Fix tests

* Fix tests

* actually fix lol

+++ src/zeroband/training/perf.py
@@ -44,7 +44,7 @@ class PerfCounter:
             return None
         return 100 * self.num_flop_per_token * tokens_per_second / self.gpu_peak_flops / self._world.world_size
 
-    def _get_peak_flops(self, device_name: str) -> int:
+    def _get_peak_flops(self, device_name: str) -> float:
         """Peak BF16 MatMul FLOPs"""
         if "A100" in device_name:
             # https://www.nvidia.com/en-us/data-center/a100/
